requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-redis==5.4.0
uvicorn==0.32.0
channels==4.1.0
channels-redis==4.2.0
whitenoise==6.8.2
psycopg2==2.9.10
environ==0.9.0